requirements.txt
openai>=1.3.3, <2.0.0
python_dotenv>=1.0.0, <2.0.0
chromadb
